added update of last generated serial number. minor fixes and improvements

+++ Contracts/GeneratorResponse.cs
@@ -3,7 +3,7 @@
 public class GeneratorResponse
 {
     public int WorkOrdersCount { get; set; }
-    public int SerialNumbersGeneratedCount { get; set; }
+    public decimal SerialNumbersGeneratedCount { get; set; }
     public string LastGeneratedSerialNumber { get; set; } = string.Empty;
     public string FilePath { get; set; } = Environment.GetFolderPath(Environment.SpecialFolder.Desktop);
 }
